Move log.py to src/
